requirements.txt
asgiref==3.7.2
Django==4.2
django-cuser==2017.3.16
django-jalali-date==1.0.2
django-modeltranslation==0.18.11
django-render-partial==0.4
django-tinymce==3.6.1
jdatetime==4.1.1
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
